requirements.txt
discord.py
requests
pytz
flask
flask-session
python-dotenv